Refactor the metrics initialization.
