+++ src/Exceptionless.Core/Properties/AssemblyInfo.cs
@@ -2,3 +2,5 @@ using System;
 using System.Runtime.CompilerServices;
 
 [assembly: InternalsVisibleTo("Exceptionless.Tests")]
+[assembly: InternalsVisibleTo("Exceptionless.Web")]
+[assembly: InternalsVisibleTo("Exceptionless.Insulation")]
