+++ src/Exceptionless.Core/Utility/IMetricsConnectionString.cs
@@ -1,0 +1,6 @@
+﻿using System;
+
+namespace Exceptionless.Core.Utility {
+    public interface IMetricsConnectionString {
+    }
+}
