+++ src/Exceptionless.Insulation/Metrics/PrometheusMetricsConnectionString.cs
@@ -1,0 +1,6 @@
+﻿using System;
+using Exceptionless.Core.Utility;
+
+namespace Exceptionless.Insulation.Metrics {
+    public class PrometheusMetricsConnectionString : IMetricsConnectionString { }
+}
